Move fake-github-middleware from server to tools-esm
